% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumall.R, R/cumsum_threshold.R
\name{cumall}
\alias{cumall}
\alias{cumany}
\alias{cumsum_threshold}
\title{Cummulative all and any}
\usage{
cumall(x)

cumany(x)

cumsum_threshold(x, threshold)
}
\arguments{
\item{x}{a \code{logical} vector.}
}
\value{
a \code{logical} vector
}
\description{
Cummulative all and any
}
\examples{
cumall(c(TRUE, TRUE, NA, TRUE, FALSE))
cumany(c(FALSE, FALSE, NA, TRUE, FALSE))
}
