% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumsum_threshold.R
\name{cumsum_threshold}
\alias{cumsum_threshold}
\title{Cummulative thresholds for sum}
\usage{
cumsum_threshold(x, threshold)
}
\arguments{
\item{x}{a \code{numeric} vector}

\item{threshold}{a \code{numeric} scalar}
}
\description{
Check
}
