% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lest-package.R
\docType{package}
\name{lest-package}
\alias{lest}
\alias{lest-package}
\title{lest: Vectorised Nested if-else Statements Similar to SQL CASE WHEN}
\description{
A fork of the functions case_when and if_else from dplyr,
  but without any external dependencies.
}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com}

}
\keyword{internal}
