% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lest-package.R
\docType{package}
\name{lest-package}
\alias{lest}
\alias{lest-package}
\title{lest: Vectorised Conditionals Similar to SQL CASE WHEN}
\description{
This is a fork of the functions case_when and if_else from dplyr,
but without any dependencies except base
}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com}

}
\keyword{internal}
