% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lest-package.R
\docType{package}
\name{lest-package}
\alias{lest}
\alias{lest-package}
\title{lest: Vectorised Nested if-else Statements Similar to CASE WHEN in 'SQL'}
\description{
Functions for vectorised conditional recoding of variables. case_when()
  enables you to vectorise multiple if and else statements (like 'CASE WHEN'
  in 'SQL'). if_else() is a stricter and more predictable
  version of ifelse() in 'base' that preserves attributes. These functions are 
  forked from 'dplyr' with all package dependencies removed and behave
  identically to the originals.
}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com}

}
\keyword{internal}
